fix: Cards responsive behaviour on contact screen
- Create ContactHelper file to provide helper functions related to Contact
screen
- Create separate Stateless widget for Connect with us card and implemented
on contact screen
- Create separate Statefull widget for Emergency contacts card and
implemented on contact screen
- Refactor Contact screen to Stateless widget and update to adapt according
to screen width
- Minor UI corrections on contact screen, connect with us card and emergency
contacts card to provide consistent visuals
